requirements.txt
# ==========================================
# Finance Analyst Skill - Python Dependencies
# ==========================================
# 用法：
#   pip install -r requirements.txt
# 
# 或者只安裝特定的套件：
#   pip install pandas numpy yfinance

# ==========================================
# 1️⃣ 數據處理和獲取
# ==========================================
pandas>=2.0.0              # 資料處理、DataFrame
numpy>=1.24.0              # 數值計算、陣列運算
yfinance>=0.2.30           # Yahoo Finance API（股票數據）
requests>=2.31.0           # HTTP 請求（通用工具）
beautifulsoup4>=4.12.0     # 網頁爬蟲（備用）

# ==========================================
# 2️⃣ 技術指標計算
# ==========================================
pandas-ta>=0.3.14          # 技術分析套件（RSI、MACD 等）
ta>=0.10.2                 # 另一個技術分析庫（備用）

# ==========================================
# 3️⃣ 統計和機率分析
# ==========================================
scipy>=1.10.0              # 科學計算（統計、最佳化）
scikit-learn>=1.3.0        # 機器學習（相似情況查詢）
statsmodels>=0.14.0        # 統計模型

# ==========================================
# 4️⃣ 設定管理
# ==========================================
pyyaml>=6.0                # 讀取 YAML 設定檔（config.yaml）
python-dotenv>=1.0.0       # 讀取 .env 環境變數

# ==========================================
# 5️⃣ Claude AI 整合
# ==========================================
anthropic>=0.7.0           # Claude API（主要工具）

# ==========================================
# 6️⃣ 日誌記錄
# ==========================================
python-logging>=0.5.1.2    # 日誌系統

# ==========================================
# 7️⃣ 測試工具（開發用，可選）
# ==========================================
pytest>=7.4.0              # 單元測試框架
black>=23.0.0              # 代碼格式化
flake8>=6.0.0              # 代碼檢查
mypy>=1.5.0                # 類型檢查

# ==========================================
# 8️⃣ 數據視覺化（可選）
# ==========================================
matplotlib>=3.8.0          # 繪圖庫（基礎）
plotly>=5.17.0             # 互動圖表

